Change colors and fix price to be a float

+++ app/src/main/java/com/example/supermarket/GroceryItem.kt
@@ -13,7 +13,7 @@ data class GroceryItem(
     var itemQuantity: Int,
 
     @ColumnInfo(name = "itemPrice")
-    var itemPrice: Int,
+    var itemPrice: Float,
 ) {
 
     @PrimaryKey(autoGenerate = true)
